requirements.txt
aiogram==3.13.1
asyncpg==0.29.0
python-dotenv==1.0.1
aiocron==1.8 